fix config in stats
